requirements.txt
openai==0.27.7
python-dotenv==1.0.0
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
toml==0.10.2
tenacity==8.2.2
PyJWT==2.8.0
jupyter==1.0.0

